refactor(app.js): Extract /me page
